+++ frontend/src/components/LiveCollection/LiveList/LiveSlider/LiveCard/LiveDetail/index.jsx
@@ -1,0 +1,1 @@
+export { default } from './LiveDetail';
