+++ frontend/src/components/LiveCollection/LiveList/LiveSlider/LiveCard/LiveThumbnail/LiveThumbnail.jsx
