+++ frontend/src/components/LiveCollection/LiveList/LiveSlider/LiveCard/index.jsx
@@ -1,0 +1,1 @@
+export { default } from './LiveCard';
