+++ frontend/src/components/LiveCollection/LiveList/LiveSlider/index.jsx
@@ -1,0 +1,1 @@
+export { default } from './LiveSlider';
