+++ frontend/src/components/LiveCollection/LiveList/index.jsx
@@ -1,0 +1,1 @@
+export { default } from './LiveList';
